requirements.txt
bcrypt==4.0.1
beautifulsoup4==4.11.1
blinker==1.5
click==8.1.3
colorama==0.4.6
distlib==0.3.6
dnspython==2.3.0
email-validator==1.3.1
filelock==3.9.0
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.0.1
Jinja2==3.1.2
MarkupSafe==2.1.2
Pillow==9.4.0
platformdirs==2.6.2
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.1
typing_extensions==4.4.0
virtualenv==20.17.1
Werkzeug==2.2.2
WTForms==3.0.1
gunicorn==20.1.0
